grpc/schema: factor out schema related logic from bigquery
In order to opt-in for schema support, a derived
class must:

 * include grpc_dest_schema_option in the grammar
 * implement the string -> descriptor type mapping
 * store and init a Schema instance in the driver
 * implement the get_schema() method to return
   the stored Schema instance

The derived class can interface with the Schema
by these functions:
  * init()
  * empty()
  * format()
  * get_schema_descriptor()
