+++ modules/grpc/bigquery/bigquery-dest.h
@@ -37,9 +37,6 @@ void bigquery_dd_set_project(LogDriver *d, const gchar *project);
 void bigquery_dd_set_dataset(LogDriver *d, const gchar *dataset);
 void bigquery_dd_set_table(LogDriver *d, const gchar *table);
 
-gboolean bigquery_dd_add_field(LogDriver *d, const gchar *name, const gchar *type, LogTemplate *value);
-void bigquery_dd_set_protobuf_schema(LogDriver *d, const gchar *proto_path, GList *values);
-
 #include "compat/cpp-end.h"
 
 #endif
