+++ modules/grpc/common/grpc-dest.h
@@ -43,6 +43,9 @@ void grpc_dd_set_keepalive_max_pings(LogDriver *s, gint p);
 void grpc_dd_add_int_channel_arg(LogDriver *s, const gchar *name, glong value);
 void grpc_dd_add_string_channel_arg(LogDriver *s, const gchar *name, const gchar *value);
 gboolean grpc_dd_add_header(LogDriver *s, const gchar *name, LogTemplate *value);
+gboolean grpc_dd_add_schema_field(LogDriver *d, const gchar *name, const gchar *type, LogTemplate *value);
+void grpc_dd_set_protobuf_schema(LogDriver *d, const gchar *proto_path, GList *values);
+
 LogTemplateOptions *grpc_dd_get_template_options(LogDriver *d);
 GrpcClientCredentialsBuilderW *grpc_dd_get_credentials_builder(LogDriver *s);
