+++ modules/grpc/common/schema/grpc-schema.hpp
@@ -29,7 +29,18 @@
 
 #include "compat/cpp-start.h"
 #include "template/templates.h"
+#include "logpipe.h"
 #include "compat/cpp-end.h"
+
+#include <google/protobuf/descriptor.h>
+#include <google/protobuf/descriptor.pb.h>
+#include <google/protobuf/dynamic_message.h>
+#include <google/protobuf/message.h>
+#include <google/protobuf/compiler/importer.h>
+#include <grpc++/grpc++.h>
+
+#include <memory>
+#include <vector>
 
 namespace syslogng {
 namespace grpc {
@@ -61,8 +72,100 @@ struct NameValueTemplatePair
 
 };
 
+struct Field
+{
+  NameValueTemplatePair nv;
+  google::protobuf::FieldDescriptorProto::Type type;
+  const google::protobuf::FieldDescriptor *field_desc;
+
+  Field(std::string name_, google::protobuf::FieldDescriptorProto::Type type_, LogTemplate *value_)
+    : nv(name_, value_), type(type_), field_desc(nullptr) {}
+
+  Field(const Field &a)
+    : nv(a.nv), type(a.type), field_desc(a.field_desc) {}
+
+  Field &operator=(const Field &a)
+  {
+    nv = a.nv;
+    type = a.type;
+    field_desc = a.field_desc;
+
+    return *this;
+  }
+
+};
+
 class Schema
 {
+private:
+  struct Slice
+  {
+    const char *str;
+    std::size_t len;
+  };
+
+public:
+  using MapTypeFn =
+    std::function<bool (const std::string &type_in, google::protobuf::FieldDescriptorProto::Type &type_out)>;
+
+public:
+  Schema(int proto_version, const std::string &file_descriptor_proto_name, const std::string &descriptor_proto_name,
+         MapTypeFn map_type, LogTemplateOptions *template_options, LogPipe *log_pipe);
+  ~Schema();
+
+  bool init();
+  google::protobuf::Message *format(LogMessage *msg, gint seq_num) const;
+
+  bool empty() const
+  {
+    return this->fields.empty();
+  }
+
+  const google::protobuf::Descriptor &get_schema_descriptor() const
+  {
+    return *this->schema_descriptor;
+  }
+
+  /* For grammar. */
+  bool add_field(std::string name, std::string type, LogTemplate *value);
+  void set_protobuf_schema(std::string proto_path, GList *values);
+
+private:
+  void construct_schema_prototype();
+  bool load_protobuf_schema();
+  Slice format_template(LogTemplate *tmpl, LogMessage *msg, GString *value, LogMessageValueType *type,
+                        gint seq_num) const;
+  bool insert_field(const google::protobuf::Reflection *reflection, const Field &field, gint seq_num,
+                    LogMessage *msg, google::protobuf::Message *message) const;
+
+private:
+  LogPipe *log_pipe;
+  MapTypeFn map_type;
+  LogTemplateOptions *template_options;
+
+  std::string syntax;
+  std::string file_descriptor_proto_name;
+  std::string descriptor_proto_name;
+
+  struct
+  {
+    std::string proto_path;
+    GList *values = nullptr;
+
+    std::unique_ptr<google::protobuf::compiler::DiskSourceTree> src_tree;
+    std::unique_ptr<google::protobuf::compiler::MultiFileErrorCollector> error_coll;
+    std::unique_ptr<google::protobuf::compiler::Importer> importer;
+    bool loaded = false;
+  } protobuf_schema;
+
+  std::vector<Field> fields;
+
+  google::protobuf::DescriptorPool descriptor_pool;
+
+  /* A given descriptor_pool/importer instance should outlive msg_factory, as msg_factory caches prototypes */
+  std::unique_ptr<google::protobuf::DynamicMessageFactory> msg_factory;
+  const google::protobuf::Descriptor *schema_descriptor = nullptr;
+  const google::protobuf::Message *schema_prototype  = nullptr;
 };
 
 }
